Finish first pass of findbb and fix name conflict
I used a hash map to look up the target of a branch/jump instruction
in `ins_arr`. The size of a hash map is set to double of the next power
of 2 of `ins_arr_sz` to keep the load factor always less equal than 0.5
but greater than 0.25. Plus, I found that there are naming conflict
accross code_analyzer.c and riscv_parser.c when linking, which is due to
var `p` in both files.