Finish findbb function and use static functions
Now the findbb function can read the INS* array, find all leaders in it,
split it bb by bb, finally put them together in a list of bb. The size
of each bb in the list is stored in an array of int called ret_bb_sz.
The total number of bb's is stored in ret_sz.